Attempting to reduce memory leakage/bloating by using unique_ptr, continuing to work on HeightNode class, cleaning up low-priority errors elsewhere, fixing template errors.

+++ DiamondDogs/engine/renderer/core/Instance.cpp
@@ -176,7 +176,7 @@ namespace vulpes {
 
 		io.ImeWindowHandle = glfwGetWin32Window(Window);
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
-		io.DisplaySize = ImVec2(DEFAULT_WIDTH, DEFAULT_HEIGHT);
+		io.DisplaySize = ImVec2(static_cast<float>(DEFAULT_WIDTH), static_cast<float>(DEFAULT_HEIGHT));
 	}
 
 	void InstanceGLFW::SetupSurface(){
